requirements.txt
# Core dependencies (from setup.py)
typer>=0.9.0
mlflow>=2.9.0
PyYAML>=6.0
torch>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
dvc>=3.30.0

# ML dependencies
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Development dependencies
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Logging and monitoring
loguru>=0.7.2
python-dotenv>=1.0.0

# API and serving
pydantic>=2.5.0
python-multipart>=0.0.6